CH8-4
Custom videoPlayer
make controller

+++ controllers/videoController.js
@@ -49,7 +49,7 @@ export const videoDetail = async(req, res) =>{
     }=req;
     try{
         const video = await Video.findById(id).populate("creator");
-        console.log(video);
+        //console.log(video);
         res.render("videoDetail",{pageTitle : video.title,video});
     } catch(error) {
         res.redirect(routes.home);
@@ -64,7 +64,12 @@ export const getEditVideo = async(req, res) => {
     }=req;
     try{
         const video = await Video.findById(id);
-        res.render("editVideo",{pageTitle : `Edit ${video.title}`,video});
+        //video.creator -type object, req.user.id type string , so error occur
+        if(video.creator.toString() !== req.user.id.toString()){
+            throw Error();
+        }else{
+            res.render("editVideo",{pageTitle : `Edit ${video.title}`,video});
+        }
     } catch(error){
         res.redirect(routes.home);
     }
@@ -88,8 +93,15 @@ export const deleteVideo = async(req, res) => {
     const{
         params:{id}
     } =req;
+    
     try{
-      await Video.findOneAndRemove({_id:id});
+        const video = await Video.findById(id);
+        if(video.creator.toString() !== req.user.id.toString()){
+            throw Error();
+        }else{
+            await Video.findOneAndRemove({_id:id});
+        }
+
     }catch(error){
        console.log(error);
     };
